fix: generate correct addi for sp

+++ tool/sysyc_test.sh
@@ -39,7 +39,7 @@ if [ $? != 0 ]; then
     exit 1
 fi
 
-if [ $emit_llvm ]; then
+if [ $emit_llvm == true ]; then
     
     clang -Wno-override-module $sysyc_out ./test/lib/sylib.c -o $out_path/$case_name
 
@@ -56,11 +56,18 @@ else
 
     riscv64-linux-gnu-gcc $sysyc_out ./test/lib/sylib.c -o $out_path/$case_name
 
+    if [ $? != 0 ]; then
+        echo -e "\033[31m$case_full failed\033[0m"
+        echo 'riscv64-linux-gnu-gcc compile error'
+        echo "see sysyc output file $sysyc_out"
+        exit 1
+    fi
+
     docker run --rm --platform=linux/riscv64 \
         -v $(realpath $out_path):/test/out \
         -v $(realpath $case_path):/test/case \
         -v $(pwd)/tool:/test/tool \
-        debain:unstable \
+        debian:unstable \
         /bin/bash -c "/test/tool/bin_test.sh /test/case /test/out $case_name" 
 
 fi
